+++ bink/callbacks/printer.py
@@ -11,27 +11,28 @@ class ConsolePrinter(Callback):
         super().__init__()
         self.validation_label = validation_label_letter
 
-    def on_step_training(self, state):
-        epoch_str = '{:d}/{:d}(t): '.format(state[bink.EPOCH] + 1, state[bink.MAX_EPOCHS])
-        batch_str = '{:d}/{:d} '.format(state[bink.BATCH], state[bink.TRAIN_STEPS])
+    def _step(self, state, letter, steps):
+        epoch_str = '{:d}/{:d}({:s}): '.format(state[bink.EPOCH], state[bink.MAX_EPOCHS], letter)
+        batch_str = '{:d}/{:d} '.format(state[bink.BATCH], steps)
         stats_str = ', '.join(['{0}:{1:.03g}'.format(key, value) for (key, value) in state[bink.METRICS].items()])
         print('\r' + epoch_str + batch_str + stats_str, end='')
+
+    def _end(self, state, letter):
+        epoch_str = '{:d}/{:d}({:s}): '.format(state[bink.EPOCH], state[bink.MAX_EPOCHS], letter)
+        stats_str = ', '.join(['{0}:{1:.03g}'.format(key, value) for (key, value) in state[bink.METRICS].items()])
+        print('\r' + epoch_str + stats_str)
+
+    def on_step_training(self, state):
+        self._step(state, 't', state[bink.TRAIN_STEPS])
 
     def on_end_training(self, state):
-        epoch_str = '{:d}/{:d}(t): '.format(state[bink.EPOCH] + 1, state[bink.MAX_EPOCHS])
-        stats_str = ', '.join(['{0}:{1:.03g}'.format(key, value) for (key, value) in state[bink.METRICS].items()])
-        print('\r' + epoch_str + stats_str)
+        self._end(state, 't')
 
     def on_step_validation(self, state):
-        epoch_str = '{:d}/{:d}({:s}): '.format(state[bink.EPOCH] + 1, state[bink.MAX_EPOCHS], self.validation_label)
-        batch_str = '{:d}/{:d} '.format(state[bink.BATCH], state[bink.VALIDATION_STEPS])
-        stats_str = ', '.join(['{0}:{1:.03g}'.format(key, value) for (key, value) in state[bink.METRICS].items()])
-        print('\r' + epoch_str + batch_str + stats_str, end='')
+        self._step(state, self.validation_label, state[bink.VALIDATION_STEPS])
 
     def on_end_validation(self, state):
-        epoch_str = '{:d}/{:d}({:s}): '.format(state[bink.EPOCH] + 1, state[bink.MAX_EPOCHS], self.validation_label)
-        stats_str = ', '.join(['{0}:{1:.03g}'.format(key, value) for (key, value) in state[bink.METRICS].items()])
-        print('\r' + epoch_str + stats_str)
+        self._end(state, self.validation_label)
 
 
 class Tqdm(Callback):
@@ -47,14 +48,25 @@ class Tqdm(Callback):
         self._loader = None
         self.validation_label = validation_label_letter
 
+    def _on_start(self, state, letter, steps):
+        bar_desc = '{:d}/{:d}({:s})'.format(state[bink.EPOCH], state[bink.MAX_EPOCHS], letter)
+        self._loader = tqdm(total=steps, desc=bar_desc)
+
+    def _update(self, state):
+        self._loader.update(1)
+        self._loader.set_postfix(state[bink.METRICS])
+
+    def _close(self, state):
+        self._loader.set_postfix(state[bink.METRICS])
+        self._loader.close()
+
     def on_start_training(self, state):
         """Initialise the TQDM bar for this training phase.
 
         :param state: The Model state
         :type state: dict
         """
-        bar_desc = '{:d}/{:d}(t)'.format(state[bink.EPOCH] + 1, state[bink.MAX_EPOCHS])
-        self._loader = tqdm(total=state[bink.TRAIN_STEPS], desc=bar_desc)
+        self._on_start(state, 't', state[bink.TRAIN_STEPS])
 
     def on_step_training(self, state):
         """Update the bar with the metrics from this step.
@@ -62,8 +74,7 @@ class Tqdm(Callback):
         :param state: The Model state
         :type state: dict
         """
-        self._loader.update(1)
-        self._loader.set_postfix(state[bink.METRICS])
+        self._update(state)
 
     def on_end_training(self, state):
         """Update the bar with the terminal training metrics and then close.
@@ -71,8 +82,7 @@ class Tqdm(Callback):
         :param state: The Model state
         :type state: dict
         """
-        self._loader.set_postfix(state[bink.METRICS])
-        self._loader.close()
+        self._close(state)
 
     def on_start_validation(self, state):
         """Initialise the TQDM bar for this validation phase.
@@ -80,8 +90,7 @@ class Tqdm(Callback):
         :param state: The Model state
         :type state: dict
         """
-        bar_desc = '{:d}/{:d}({:s})'.format(state[bink.EPOCH] + 1, state[bink.MAX_EPOCHS], self.validation_label)
-        self._loader = tqdm(total=state[bink.VALIDATION_STEPS], desc=bar_desc)
+        self._on_start(state, self.validation_label, state[bink.VALIDATION_STEPS])
 
     def on_step_validation(self, state):
         """Update the bar with the metrics from this step.
@@ -89,8 +98,7 @@ class Tqdm(Callback):
         :param state: The Model state
         :type state: dict
         """
-        self._loader.update(1)
-        self._loader.set_postfix(state[bink.METRICS])
+        self._update(state)
 
     def on_end_validation(self, state):
         """Update the bar with the terminal validation metrics and then close.
@@ -98,5 +106,4 @@ class Tqdm(Callback):
         :param state: The Model state
         :type state: dict
         """
-        self._loader.set_postfix(state[bink.METRICS])
-        self._loader.close()
+        self._close(state)